Move user route handlers to their own controllers
