+++ controllers/add_rest_controller.js
@@ -1,0 +1,13 @@
+var Rest = require('../models/Restaurant');
+
+module.exports = function(req, res) {
+  var newRest = new Rest(req.body);
+
+  newRest.save(function(err, rest) {
+    if (err) {
+      console.log(err);
+      return res.status(500).json({msg: 'internal server error'});
+    }
+    res.status(200).json({msg: 'save successful'});
+  });
+};
