+++ controllers/sign_in_controller.js
@@ -1,0 +1,13 @@
+var User = require('../models/User');
+
+module.exports = function(req, res) {
+  req.user.generateToken(process.env.APP_SECRET, function(err, token) {
+    if (err) {
+      console.log(err);
+      return res.status(500).json({msg : 'problem generating token'});
+    }
+
+    var uriToken = encodeURIComponent(token);
+    res.redirect('/hinton/user?valid=' + uriToken);
+  });
+};
